adicao de header.css alem de arrumar o codigo para tirar os componentes do css global
