+++ src/App.css
@@ -151,7 +151,7 @@ form.custom .custom.disabled { background-color: #ddd }
 }
 .lt-ie9 form.custom div.custom.dropdown a.selector { height: 30px }
 /* The Grid ---------------------- */
-.row { width: 980px;  min-width: 768px; max-width: 100%; margin: 0 auto; }
+
 .row .row { width: auto; max-width: none; min-width: 0; margin: 0 -15px; }
 .row.collapse .column,
 .row.collapse .columns { padding: 0 }
@@ -662,53 +662,7 @@ div.progress.round { -webkit-border-radius: 1000px; -moz-border-radius: 1000px; 
 div.progress.round .meter { -webkit-border-radius: 1000px; -moz-border-radius: 1000px; -ms-border-radius: 1000px; -o-border-radius: 1000px; border-radius: 1000px; }
 
 /* MENU SUPERFISH */
-#menu-icon { display: none; /* hide menu icon initially */ }
-.sf-menu,.sf-menu li { margin: 0; padding: 0; float: right; }
-.sf-menu li { list-style: none; float: left; }
-.sf-menu a { padding: 4px 15px; display: block; color: #000; }
-.sf-menu a:hover { }
-.sf-menu ul { padding: 2px; position: absolute; display: none; /* hide dropdown */ width: 200px; }
-.sf-menu ul li { float: none; margin: 0px; padding: 0px; }
-.sf-menu li:hover > ul { display: block; /* show dropdown on hover */ }
-ul.sf-menu li li:hover ul,ul.sf-menu li li.sfHover ul { left: 200px; /* match ul width */ top: 0; }
-.sf-menu,.sf-menu * { margin: 0; padding: 0; list-style: none; }
-.sf-menu,.sf-menu * { margin: 0; padding: 0; list-style: none; }
-.sf-menu { line-height: 1.0 }
-.sf-menu ul { position: absolute; top: -999em; width: 10em; }
-.sf-menu ul li { width: 100% }
-.sf-menu li:hover { visibility: inherit }
-.sf-menu li { float: left; position: relative; margin-right: 7px; }
-.sf-menu a { display: block; position: relative; }
-.sf-menu li:hover ul, .sf-menu li.sfHover ul { left: 0%; padding-top:0px;top:40px; /* match top ul list item height */ z-index: 99; }
-ul.sf-menu li:hover li ul, ul.sf-menu li.sfHover li ul { top: -999em }
-ul.sf-menu li li:hover ul, ul.sf-menu li li.sfHover ul { left: 10em; /* match ul width */ top: 0; }
-ul.sf-menu li li:hover li ul, ul.sf-menu li li.sfHover li ul { top: -999em }
-ul.sf-menu li li li:hover ul, ul.sf-menu li li li.sfHover ul { left: 10em; /* match ul width */ top: 0; }
-.sf-menu { float: right; margin-bottom: 1em; }
-.sf-menu a { padding: .95em 1.1em; text-decoration: none; }
-.sf-menu a, .sf-menu a:visited { /* visited pseudo selector so IE6 applies text colour*/ color: #444; }
-.sf-menu li { }
-.sf-menu li li { background: #fff; width: 200px; }
-.sf-menu li li li { background: #fff }
-.sf-menu li:hover,.sf-menu li.sfHover,.sf-menu a:focus,.sf-menu a:hover,.sf-menu a:active { outline: 0; color: #ccc; }
-.sf-menu a.sf-with-ul { min-width: 1px }
-.sf-sub-indicator { position: absolute; display: block; right: .75em; top: 1.05em; /* IE6 only */ width: 10px; height: 10px; text-indent: -999em; overflow: hidden; }
-a > .sf-sub-indicator { /* give all except IE6 the correct values */ top: 0.9em; background-position: 0 -100px; }
-#nav-wrap { margin-top: 30px }
-a:focus > .sf-sub-indicator,
-a:hover > .sf-sub-indicator,
-a:active > .sf-sub-indicator,
-li:hover > a > .sf-sub-indicator,
-li.sfHover > a > .sf-sub-indicator { background-position: -10px -100px }
-.sf-menu ul .sf-sub-indicator { background-position: -10px 0 }
-.sf-menu ul a > .sf-sub-indicator { background-position: 0 0 }
-.sf-menu ul a:focus > .sf-sub-indicator,
-.sf-menu ul a:hover > .sf-sub-indicator,
-.sf-menu ul a:active > .sf-sub-indicator,
-.sf-menu ul li:hover > a > .sf-sub-indicator,
-.sf-menu ul li.sfHover > a > .sf-sub-indicator { background-position: -10px 0 }
-.sf-shadow ul { padding: 0 8px 9px 0; border-top: 2px solid #33877D; width: 200px; }
-.sf-shadow ul.sf-shadow-off { background: transparent; }
+
 ul.flyout, .nav-bar li ul {
   list-style: none;
   padding: 0;
