+++ src/components/header/header.styles.css
@@ -1,0 +1,94 @@
+
+.row {
+
+    width: 980px;
+    min-width: 768px;
+    max-width: 100%;
+    margin: 0 auto;
+
+}
+
+.columns { 
+    float: left;
+    min-height: 1px;
+    padding: 0 15px; 
+    position: relative; 
+}
+
+
+.logo {
+    margin-top: 34px;
+    margin-bottom: 30px;
+}
+
+.nav-wrap {
+    margin-top: 30px
+}
+
+.sf-menu li li {
+    background: #fff;
+    width: 200px;
+}
+
+
+.sf-menu li:hover ul, .sf-menu li.sfHover ul {
+    left: 0%;
+    padding-top:0px;
+    top:40px; /* match top ul list item height */ z-index: 99;
+}
+
+
+.sf-menu,.sf-menu * {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+}
+
+
+.sf-menu li:hover > ul {
+    display: block; /* show dropdown on hover */
+}
+
+.sf-menu ul {
+    padding: 2px;
+    position: absolute;
+    display: none; /* hide dropdown */
+    width: 200px;
+}
+
+.sf-menu,.sf-menu li {
+    margin: 0;
+    padding: 0;
+    float: right;
+}
+
+.sf-menu li {
+    list-style: none;
+    float: left;
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
